+++ src/components/Breadcrumbs/BreadcrumbsTypes.tsx
@@ -1,6 +1,7 @@
 export enum ColorVariants {
   Blogs = "blogs",
   BooksAndMore = "booksAndMore",
+  Education = "education",
   Locations = "locations",
   Research = "research",
   WhatsOn = "whatsOn",