+++ src/components/MultiSelect/MultiSelect.stories.tsx
@@ -81,7 +81,7 @@ export const MultiSelectStory: Story<MultiSelectProps> = (args) => {
       itemIds = selectedItems[multiSelectId].items.slice();
       // If termId exists, remove it from the array.
       if (itemIdExists) {
-        itemIds = itemIds.filter((id) => id != itemId);
+        itemIds = itemIds.filter((id) => id !== itemId);
       } else {
         // Add it to the array, but modify the copy, not the original.
         itemIds.push(itemId);
@@ -125,7 +125,7 @@ export const MultiSelectDialogStory: Story<MultiSelectDialogProps> = (args) => {
       itemIds = selectedItems[multiSelectId].items.slice();
       // If termId exists, remove it from the array.
       if (itemIdExists) {
-        itemIds = itemIds.filter((id) => id != itemId);
+        itemIds = itemIds.filter((id) => id !== itemId);
       } else {
         // Add it to the array, but modify the copy, not the original.
         itemIds.push(itemId);