+++ src/utils/componentCategories.ts
@@ -92,9 +92,10 @@ const categories = {
     components: ["Breadcrumbs", "Link", "Menu", "Pagination", "Subnavigation"],
   },
   chakra: {
-    title: "Chakra/Layout",
+    title: "Chakra/Chakra COmponent/Layout",
     components: [
       "Box",
+      "Flex",
       "Grid",
       "Center, Circle, Square",
       "Stack, HStack, VStack",